add ckfinder,ckeditor,and delete all for web..end lesson 31->34..common baby

+++ XD_WEB.WEB1/Models/ProductViewModel.cs
@@ -4,7 +4,7 @@ namespace XD_WEB.WEB1.Models
 {
     public class ProductViewModel
     {
-        public long ID { set; get; }
+        public int ID { set; get; }
 
         public string Name { set; get; }
